Add files via upload
The code works if "pyautogui" is installed in you computer
The code automatically types any number of message that is entered in the input field, in any place that cursor is. Its also able to send the messages to someone if the cursor is placed in the type box of the chat.